ARM updates
Due to the ABI we are using we have to use the earm architecture
moniker for the build system to behave correctly. This involves
then some headers to move around.

There is also a few related Makefile updates as well as minor
source code corrections.

+++ sys/arch/arm/include/int_types.h
@@ -62,7 +62,7 @@ typedef	unsigned long long int __uint64_t;
 
 /* 7.18.1.4 Integer types capable of holding object pointers */
 
-typedef long int	       __intptr_t;
-typedef unsigned long int     __uintptr_t;
+typedef	int		       __intptr_t;
+typedef	unsigned int	      __uintptr_t;
 
 #endif	/* !_ARM32_INT_TYPES_H_ */
